slices FE refactor
